fix: update import paths after folder move
